services/easydriver/requirements.txt
grpcio==1.28.1
grpcio-tools==1.28.1
protobuf==3.11.3
RPi.GPIO==0.7.0
six==1.15.0
